Move plot_summary into the main project
